unify parallel config
